+++ main.py
@@ -7,13 +7,15 @@ from kivy.app import App
 from kivy.properties import ListProperty
 
 from cal_functions import CoreFunctions
-import palette
+from palette import RED, background_colour, text_colour, selected_colour, disabled_colour
 
 # kivy doesn't require def __init__() - uses build() instead
 class MainApp(App):
-    colour_light_base = ListProperty(palette.LIGHT_BACKGROUND)
-    colour_light_text = ListProperty(palette.LIGHT_TEXT)
-    colour_light_blue = ListProperty(palette.LIGHT_BLUE)
+    colour_background = ListProperty(background_colour)
+    colour_text = ListProperty(text_colour)
+    colour_selected = ListProperty(selected_colour)
+    colour_disabled = ListProperty(disabled_colour)
+    colour_holiday = ListProperty(RED)
     def build(self) -> CoreFunctions:
         self.title = 'Calendar App'
         self.locale = 'en_US.UTF-8'